cmd/k8s-operator: optionally update dnsconfig Configmap with DNS records for proxies.
If DNSConfig has been applied to cluster,
update the dnsconfig ConfigMap that the nameserver mounts
and reads records from with records for MagicDNS names to cluster IP addresses.

The records are created as follows:
* For cluster ingress proxies there will be
a record mapping the MagicDNS name of the proxy
device and each proxy Pod's IP address.
* For cluster egress proxies there will be
a record for each proxy Pod, mapping
the MagicDNS name of the exposed
tailnet workload to the proxy Pod's IP.

Currently only one Pod exists per proxy,
but HA is planned for future and this mechanism
will support that.

Updates tailscale/tailscale#10499

Signed-off-by: Irbe Krumina <irbe@tailscale.com>